Created the Hand class in place of a list of Cards in the player class. You can add and remove cards from the hand.
Modified the Player to now be an abstract class instead of an interface, as method implementations were similar in both ComputerPlayer and HumanPlayer. Player contains the implemented methods for similar methods in both subclasses except for putCard() in ComputerPlayer, which extends the behaviour (selects a random card and call superclass).

Modified Game to now contain a change turn method (changeCurrentTurn). This method is not implemented as I believe there is a possible design pattern we could use here (iterator) that needs further discussion. Updated playCard() so that it takes the former currentCard, adds it to the gameDeck, and takes the inputted card as the new currentCard.

Modified Deck so that it contains add and remove card methods. However, this still needs a constructor, which will be updated later.