Add GameManager and related interfaces
Part of the reworking of how interactors interact with the game; currently they reference a game instance directly, however this causes trouble when the PlayerGameInteractor, which handles game logic, needs an instance to function, and one isn't created in time by the GameCreationInteractor. Having a game proxy class that encompasses both these use cases in the form of interfaces aims to solve this problem.